requirements.txt
streamlit==1.38.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1

# ML / EDA
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
